refactor(songs): extract search/filter controls into useSongFilters
Symmetric to the useSongFeed (data) extraction: move the debounced
search state, the grade/class filter state, the class-options fetch, and
the derived grade/class option lists out of SongsPage into a
useSongFilters hook. It returns a ready-made filters object for
useSongFeed plus the state/derived lists the toolbar renders, so the
page is two hooks (filters + feed) and presentational JSX.

Selecting a grade clearing the class is now one selectGrade() instead of
an inline two-setter onChange. Behaviour unchanged.
